+++ CHANGELOG.md
@@ -41,6 +41,7 @@ This project adheres to [Semantic Versioning](https://semver.org).
 		that contains a single comment.
 	* Change ``ID3v2Frames`` to present a list of comments for ``comment`` key.
 * Rename ``formats.vorbis`` module to ``formats.vorbis_comments``.
+* Load most commonly used unoffical ID3v2 frames.
 
 ### Removed